Kermis v1

+++ attracties/interfaces/GokAttractie.java
@@ -2,5 +2,5 @@ package com.example.kermis.attracties.interfaces;
 
 public interface GokAttractie {
 
-    int kansSpelBelastingBetalen();
+    double kansSpelBelastingBetalen();
 }
